+++ test/mocks/schema_mocks.dart
@@ -3,7 +3,8 @@ import 'package:jar/jar.dart';
 
 class MockStringSchema extends Mock implements JarSchema<String, MockStringSchema> {
   @override
-  JarResult validate(String? value, [Map<String, dynamic>? values]) {
+  JarResult validate(String? value,
+      [Map<String, dynamic>? values, String? path, Object? root]) {
     return JarResult.success();
   }
 }
@@ -14,7 +15,8 @@ class MockFailingStringSchema extends Mock implements JarSchema<String, MockFail
   MockFailingStringSchema(this.errorMessage);
   
   @override
-  JarResult validate(String? value, [Map<String, dynamic>? values]) {
+  JarResult validate(String? value,
+      [Map<String, dynamic>? values, String? path, Object? root]) {
     return JarResult.error(errorMessage);
   }
 }